requirements.txt
certifi==2022.9.24
babel==2.10.3
flask-babel==2.0.0
flask==2.2.2
jinja2==3.1.2
lxml==4.9.1
pygments==2.13.0
python-dateutil==2.8.2
pyyaml==6.0
httpx[http2]==0.21.2
Brotli==1.0.9
uvloop==0.17.0
httpx-socks[asyncio]==0.7.2
langdetect==1.0.9
setproctitle==1.3.2
redis==4.3.4
markdown-it-py==2.1.0
typing_extensions==4.3.0
pygeonlp
feedparser
psycopg2
aiohttp
tzdata
pycrypto
tldextract
